requirements.txt
absl-py==0.9.0
amqp==2.5.2
asgiref==3.2.3
astor==0.8.1
astroid==2.3.3
autopep8==1.4.4
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cycler==0.10.0
Django==3.0.2
falcon==2.0.0
falcon-cors==1.1.7
gast==0.3.3
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
image==1.5.20
importlib-metadata==1.5.0
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.1.2.30
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.2
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.1
vine==1.3.0
Werkzeug==0.16.1
wrapt==1.11.2
zipp==2.2.0
